Map and enemy intialization, hud creation
Made the first map, I tried to incorporate the navmesh and ai navigation but I couldn't get the mush to align with the path. I stuck with the checkpoints for now. Added billboard affect above castle for health bar. Created a canvas that overlays the camera. It has 3 sliders and a confirm button. Needs to be better sized, bigger.